WIP New memory allocator
The previous memory allocator turned out to be slow in some circumstances,
expecially for operations that performed large numbers of small memory
allocations. This introduces a new default memory allocator for the asmjs
compile target.

It is currently WIP because there is an outstanding issue with the approach:
blocks of freed memory are only compacted if they are immediately adjacent
and the block being freed is the first of the two blocks. That means that
if one free block of memory precedes another block of memory, the two
will never be compacted togheter unless the first block is allocated and
freed before the second can be freed and allocated again.